# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PublicProblemResponse": {
        "description": "Typed-client-safe problem fields guaranteed across every API error.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "PublicProblemResponse",
        "type": "object"
      },
...
      "TravelDetailsRequest": {
        "additionalProperties": false,
        "properties": {
          "destination_summary": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Summary"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "title": "TravelDetailsRequest",
        "type": "object"
      },
      "TravelResponse": {
        "properties": {
          "destination_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Summary"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "plan_id": {
            "format": "uuid",
            "title": "Plan Id",
            "type": "string"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/TravelSegmentResponse"
            },
            "title": "Segments",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "plan_id",
          "destination_summary",
          "notes",
          "version",
          "segments"
        ],
        "title": "TravelResponse",
        "type": "object"
      },
      "TravelSegmentRequest": {
        "additionalProperties": false,
        "description": "Local wall-clock times with IANA zones (flights cross zones), or dates only.",
        "properties": {
          "arrival_local": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrival Local"
          },
          "arrival_timezone": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrival Timezone"
          },
          "departure_local": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Departure Local"
          },
          "departure_timezone": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Departure Timezone"
          },
          "destination_label": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Label"
          },
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "origin_label": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Label"
          },
          "segment_type": {
            "enum": [
              "flight",
              "train",
              "bus",
              "car",
              "ferry",
              "lodging",
              "other"
            ],
            "title": "Segment Type",
            "type": "string"
          },
          "sort_order": {
            "default": 0,
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Sort Order",
            "type": "integer"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "timing_mode": {
            "enum": [
              "date",
              "datetime"
            ],
            "title": "Timing Mode",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "segment_type",
          "timing_mode"
        ],
        "title": "TravelSegmentRequest",
        "type": "object"
      },
      "TravelSegmentResponse": {
        "properties": {
          "arrival_local": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrival Local"
          },
          "arrival_timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrival Timezone"
          },
          "arrives_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrives At"
          },
          "departs_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Departs At"
          },
          "departure_local": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Departure Local"
          },
          "departure_timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Departure Timezone"
          },
          "destination_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Label"
          },
          "end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "origin_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Label"
          },
          "segment_type": {
            "title": "Segment Type",
            "type": "string"
          },
          "sort_order": {
            "title": "Sort Order",
            "type": "integer"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "timing_mode": {
            "enum": [
              "date",
              "datetime"
            ],
            "title": "Timing Mode",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "segment_type",
          "title",
          "origin_label",
          "destination_label",
          "timing_mode",
          "start_date",
          "end_date",
          "departure_local",
          "departure_timezone",
          "departs_at",
          "arrival_local",
          "arrival_timezone",
          "arrives_at",
          "sort_order",
          "version"
        ],
        "title": "TravelSegmentResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API for plans with your people.",
    "title": "Beluno API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/plans/{plan_id}/travel": {
      "get": {
        "operationId": "get_travel_v1_plans__plan_id__travel_get",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TravelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Bearer authentication is missing, invalid, expired, or revoked."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The resource is absent or not visible to the caller."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "A dependency or feature is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Travel",
        "tags": [
          "travel"
        ]
      },
      "put": {
        "description": "Create the travel extension, or replace it with the current ``If-Match`` version.",
        "operationId": "put_travel_details_v1_plans__plan_id__travel_put",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TravelDetailsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TravelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Bearer authentication is missing, invalid, expired, or revoked."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The current relationship does not allow this action."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The resource is absent or not visible to the caller."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request conflicts with the current resource state."
          },
          "412": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The If-Match version is stale."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request does not satisfy the contract."
          },
          "428": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "An If-Match header is required."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "A dependency or feature is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Put Travel Details",
        "tags": [
          "travel"
        ]
      }
    },
    "/v1/plans/{plan_id}/travel/segments": {
      "post": {
        "operationId": "add_segment_v1_plans__plan_id__travel_segments_post",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TravelSegmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TravelSegmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Bearer authentication is missing, invalid, expired, or revoked."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The current relationship does not allow this action."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The resource is absent or not visible to the caller."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request conflicts with the current resource state."
          },
          "412": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The If-Match version is stale."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request does not satisfy the contract."
          },
          "428": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "An If-Match header is required."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "A dependency or feature is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Segment",
        "tags": [
          "travel"
        ]
      }
    },
    "/v1/plans/{plan_id}/travel/segments/{segment_id}": {
      "delete": {
        "operationId": "delete_segment_v1_plans__plan_id__travel_segments__segment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "segment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Segment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Bearer authentication is missing, invalid, expired, or revoked."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The current relationship does not allow this action."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The resource is absent or not visible to the caller."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request conflicts with the current resource state."
          },
          "412": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The If-Match version is stale."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request does not satisfy the contract."
          },
          "428": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "An If-Match header is required."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "A dependency or feature is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Segment",
        "tags": [
          "travel"
        ]
      },
      "put": {
        "operationId": "update_segment_v1_plans__plan_id__travel_segments__segment_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "segment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Segment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TravelSegmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TravelSegmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "Bearer authentication is missing, invalid, expired, or revoked."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The current relationship does not allow this action."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The resource is absent or not visible to the caller."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request conflicts with the current resource state."
          },
          "412": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The If-Match version is stale."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "The request does not satisfy the contract."
          },
          "428": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "An If-Match header is required."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProblemResponse"
                }
              }
            },
            "description": "A dependency or feature is temporarily unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Segment",
        "tags": [
          "travel"
        ]
      }
    }
  }
}